zsh: Remove --short from git status
Too used to the verbose format and I find it easier to read
in general, especially with submodules and many deletions/additions.

The short format uses a single character to describe the change which
gets confusing when its all in the same green meaning its staged.

e.g. M and D are both green but that makes it very hard to read.